refactor(ci): extract Traefik compose diff guard
